% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataYoutube2.R
\name{CollectDataYoutube2}
\alias{CollectDataYoutube2}
\title{Collect YouTube Video Comments}
\usage{
CollectDataYoutube2(credential = NA, videoIDs, verbose, writeToFile = FALSE,
  maxComments)
}
\arguments{
\item{credential}{optional, character vector containing authentication token.}

\item{videoIDs}{character vector, specifying one or more YouTube video IDs.}

\item{verbose}{boolean, if \code{TRUE} then this function will output
information to the console during processing. Default is \code{FALSE}.}

\item{writeToFile}{boolean, if \code{TRUE} then the data is saved to file in
current working directory in \code{CSV} format. The file name will be
the current system time and data type as 
\code{MMM_DD_HH_MM_SS_YYYY_YoutubeData2.csv}. The file is appended as each
videos comments data is collected. Default is \code{FALSE}.}

\item{maxComments}{numeric integer, sets limit for how many total comments
to collect. This is a very rough measure that only works if comments from
multiple videos are requested. If \code{maxComments} is reached in the
current video, no further videos will be processed. By default this
function attempts to collect all comments.}
}
\value{
data frame object of class \code{dataSource.youtube2}
}
\description{
YouTube Video Comments Collection with R tuber package
DO NOT CALL DIRECTLY - USE \code{Collect}
}
\details{
Collection performed using the Google API via the tuber package:
http://soodoku.github.io/tuber/reference/get_all_comments.html

Collects and structures the data into a data frame of class
\code{dataSource.youtube2} suitable for creating networks such as
a unimodal actor network (\code{CreateActorNetwork}).
}
\seealso{
\code{AuthenticateWithYoutube2} must be run first or no data will
be collected.
}
