% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateActorNetwork.youtube2.R
\name{CreateActorNetwork.youtube2}
\alias{CreateActorNetwork.youtube2}
\title{Create YouTube Actor Network}
\usage{
\method{CreateActorNetwork}{youtube2}(commentsData, directIsolates = TRUE,
  writeToFile = FALSE)
}
\arguments{
\item{commentsData}{data frame containing comments data collected and
structured by \code{CollectDataYoutube2}.}

\item{directIsolates}{boolean, if \code{TRUE} top level comments will be directed
to the creator of the video that the actors are commenting on. If \code{FALSE}
top level comments are not directed and actors will be isolates if they have
not replied to any comments by other actors.Default is \code{TRUE}.}

\item{writeToFile}{boolean, if \code{TRUE} then igraph data is saved to a 
file in the current working directory in \code{graphml} format. The file
name will be the current system time and network type as
\code{MMM_DD_HH_MM_SS_YYYY_YoutubeActorNetwork2.graphml}. Default
is \code{FALSE}.}
}
\value{
igraph object containing the actor network
}
\description{
Creates a Unimodal Actor Network from YouTube Video Comments
DO NOT CALL DIRECTLY - USE \code{Create}
}
\details{
Creates a unimodal actor network based on comments and replies to one or
more youtube videos.
}
\note{
\code{directIsolates=TRUE} option will use additional
Google API resource points as it makes a video details request for 
each youtube video that comments were collected from. As such this
option will also require \code{AuthenticateWithYoutube2} to be
called first. Otherwise set this to \code{FALSE}.
}
\seealso{
\code{CollectDataYoutube2} must be called first to collect an
appropriately structured data frame.
}
